checkpoint(library/Base64): moves "Base64/CharacterSequence" to "Sequence/Base64"
