+++ src/library/Sequence/Base64/ConformanceError.ts
